Move Components to a stub, so we won't run into errors with PHP version below 7.4
