+++ tests/Unit/PublicPropertyManagerTest.php
@@ -3,7 +3,6 @@
 namespace Tests\Unit;
 
 use Illuminate\Support\Str;
-use Livewire\Component;
 use Livewire\Exceptions\CannotRegisterPublicPropertyWithoutImplementingWireableException;
 use Livewire\LivewirePropertyManager;
 use Livewire\Livewire;
@@ -17,6 +16,8 @@ class PublicPropertyManagerTest extends TestCase
         if (version_compare(PHP_VERSION, '7.4', '<')) {
             $this->markTestSkippede('Skip public property tests if the version is below PHP 7.4');
         }
+
+        require_once __DIR__.'/PublicPropertyManagerStubs.php';
     }
 
     /** @test */
@@ -28,7 +29,7 @@ class PublicPropertyManagerTest extends TestCase
     /** @test */
     public function it_will_throw_an_exception_if_registering_a_class_not_implementing_the_wireable_interface()
     {
-        $this->markTestSkipped(" We will throw an exception, but I guess we'll need a new interface. Until then, do nothing.");
+        $this->markTestSkipped("We will throw an exception, but I guess we'll need a new interface. Until then, do nothing.");
 
         $this->expectException(CannotRegisterPublicPropertyWithoutImplementingWireableException::class);
 
@@ -64,56 +65,5 @@ class PublicPropertyManagerTest extends TestCase
         );
 
         $this->assertCount(1, app(LivewirePropertyManager::class)->properties());
-    }
-}
-
-class CustomPublicClass
-{
-    public $message;
-
-    public function __construct($message, $embeddedMessage)
-    {
-        $this->message = $message;
-    }
-}
-
-class CustomResolverClass
-{
-    public ?CustomPublicClass $class;
-
-    public function __construct(CustomPublicClass $class) {
-        $this->class = $class;
-    }
-
-    public function toLivewire()
-    {
-        return [
-            'message' => $this->class->message,
-        ];
-    }
-
-    public static function fromLivewire($value)
-    {
-        return new CustomPublicClass($value['message'], 'embedded message which is missing right now');
-    }
-}
-
-class ComponentWithCustomPublicProperty extends Component
-{
-    public ?CustomPublicClass $wireable;
-
-    public function mount($wireable)
-    {
-        $this->wireable = $wireable;
-    }
-
-    public function removeWireable()
-    {
-        $this->wireable = null;
-    }
-
-    public function render()
-    {
-        return view('wireables');
     }
 }
